Implement print_last_digit function
Adding a Python function called 'print_last_digit' that takes an integer as input. The function calculates the last digit of the absolute value of the input number using the modulo operator (%) with 10. It then prints the last digit without a newline and returns it. This function is useful for printing and extracting the last digit of a number while avoiding line breaks.