Implement FizzBuzz function
Implementing a Python function named 'fizzbuzz' that prints numbers from 1 to 100. For multiples of 3, it prints 'Fizz', for multiples of 5, it prints 'Buzz', and for numbers that are multiples of both 3 and 5, it prints 'FizzBuzz'. This function is commonly used as a programming exercise to demonstrate conditional statements and looping.