Utilize calculator_1 module for arithmetic operations
Imported functions 'add', 'sub', 'mul', and 'div' from the 'calculator_1' module. Created two integer variables 'a' and 'b' with values 10 and 5 respectively. Utilized the imported functions to perform addition, subtraction, multiplication, and division operations on 'a' and 'b'. Printed the results of each operation in the format 'a operator b = result'. This integration allows performing various arithmetic operations using the functions from the imported module.