Print value of a variable from another module
Implemented a script that checks if it's being executed directly and not imported as a module. Imported the 'variable_load_5' module and specifically imported the variable 'a' from it. Printed the value of the variable 'a'. This script allows users to display the value of the variable 'a' from the 'variable_load_5' module.