Implement function to check divisibility by 2
Added a Python script containing the 'divisible_by_2' function. This function takes a list of integers as input and returns a new list of Boolean values, where each value corresponds to whether the corresponding integer in the input list is divisible by 2. True is appended to the result list for integers that are divisible by 2, and False is appended for non-divisible integers.