Swap values of variables using tuple unpacking
Adding a Python script that demonstrates swapping the values of two variables using tuple unpacking. The script initializes two variables 'a' and 'b' with values 89 and 10 respectively. It then swaps their values using the syntax 'a, b = b, a', effectively exchanging their values. The final values of 'a' and 'b' are printed using the 'format' method of the 'print' function.