Print list of integers using format method
Adding a Python script that demonstrates printing a list of integers using the 'format' method. The script defines a function 'print_list_integer' that takes a list of integers as an argument and iterates through the list using a 'for' loop. It uses the 'format' method to print each integer.