Find elements that are unique to each set
Implementing a Python function 'only_diff_elements' designed to identify and gather elements that are unique to each of the two input sets. The function employs the 'symmetric_difference' method provided by sets, enabling an efficient computation of the symmetric difference between the sets. This results in a new set containing elements exclusively present in either of the input sets, excluding elements shared between the two.