Add 'Node' and 'SinglyLinkedList' classes
Added the 'Node' class, which defines a node of a singly linked list, with data and next_node attributes. The 'SinglyLinkedList' class represents a singly linked list and includes methods for inserting nodes in sorted order and displaying the linked list as a string. The data attribute in the 'Node' class is now validated to be an integer, and the next_node attribute is validated to be a 'Node' object.