Implement lookup function
Implemented the 'lookup' function in the Python script. This function takes an object as input and returns a list of available attributes and methods of that object using the 'dir' function. It can be used for inspecting objects and exploring their properties. The function has been tested and is ready for use.