Implement inherits_from function
Implemented the 'inherits_from' function in the Python script. This function checks if an object is an instance of a class that inherited (directly or indirectly) from the specified class. It returns 'True' in such cases, otherwise 'False'. This function can be useful for checking class inheritance relationships in Python programs. The function has been tested and is ready for use.