Add Base class for managing id attribute
Added the Base class, which provides functionality for managing an 'id' attribute for other classes. It includes methods to convert lists of dictionaries to JSON strings and vice versa, to save instances to a JSON file, to create instances from dictionaries, and to load instances from a JSON file. The '__nb_objects' class variable is used to auto-generate 'id' values.